+++ backend/src/main/java/com/back/domain/product/product/service/ProductService.java
@@ -23,7 +23,7 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public Optional<Product> findById(int id) {
+    public Optional<Product> findById(long id) {
         return productRepository.findById(id);
     }
